Fix R.java generation with deps_query usage
Summary:
R.java generation is part of the build graph enhancement which was not
supplied with the additional results from `deps_query` expansion.
This results in R.java entries being missing for any dep added in this
fashion causing build breaks in such a way that manually adding them
to the regular declared deps list would fix it.

Test Plan:
Verified locally with a simple android_library and android_resource
pair where deps_query was used to locate the android_resource.

Reviewed By: illicitonion

fbshipit-source-id: d6263af