Extend `verify-cache` command to verify rule key cache
Summary:
Updates the `verify-cache` command to re-calculate and compare rule
keys with cached values.  The verification process is a bit unusual,
as it assumes the action graph is still valid (since it doesn't have
a target graph to use to get an action graph).

Test Plan:
Ran a build with rule key caching enabled and then ran the verification
command.  Then, modified Buck so that the rule key cache no longer
invalidates based on watchman events, re-ran the build, modified a file
and then ran the verification command again, this time getting an error.

Reviewed By: dreiss

fbshipit-source-id: 70d4dda