Remove sorting of stubs to maintain order found in original JARs
Summary:
The existing implementation of ABI generation is that it collects elements in a TreeSet to keep elements sorted. This is so that re-ordering elements such as fields or methods would result in the same ABI output, which would result in not requiring to recompile dependents.

This optimization needs to be removed due to an issue that came up in compiling against ABI jars. It is possible for annotation processors to take a dependency on the specific ordering of members.

Keeping the order to match the source code not only fixes issues with code generation, it also has a nice property of matching the original full JAR and removing any other possible errors that may occur from having things sorted.

Note: Inner classes and interfaces are the only elements that are actually scanned and written in reverse order. This is just how the behavior was discovered in analyzing how the original JARs keep their order.

Changes:
- `ClassMirror` is changed from using `TreeSet` to using `LinkedHashSet` to maintain insertion order
- `ClassVisitorDriverFromElement` collects inner classes and interfaces and scans them in reverse to match order in original JARs

Test Plan: - `StubJarTest` is updated to not check if stubs are sorted. Instead, it checks that the order of stub members matches the order in the original members. All tests pass, including a new test added.

Reviewed By: jkeljo

fbshipit-source-id: 5fcea64