Fix how the icon gets initialized for the buck tool window.
Summary:
The icon for the buck tools window  has been broken since BuckIcons
was moved out of the package "icons."  It looks like that package
is a blessed location for icons.  Rather than restructure our code to
locate the BuckIcons class exactly where IntelliJ wants it, we
remove the icon from the plugin.xml declaration (leaving it with the
default icon) and set the ToolWindow to use the special buck icon in
the constructor (where we're free to use a path of our choosing).

Test Plan: Build/install locally.

Reviewed By: JonShemitz

fbshipit-source-id: 098f029