Improved strategy to include the correct headers in the merged header map
Summary:
Now, we gather the dependencies of the main targets and include their headers
in the merged header map. It will prevent the build to include headers
from targets that are not dependencies.

For example, if we have two targets within the same BUCK file that produce
a header log.h and source file includes that header and only one of those
targets is used, we'd like to use the log.h header from that target.

Test Plan:
- have a BUCK file with two targets A and B
- A has an exported A/log.h -> exported/log.h file
- B has an exported B/log.h -> exported/log.h file but different
- have a main target that depend only on A
- generating the merged header map should only add A/log.h and not B/log.h

Before the diff, the header map would also include B/log.h

Reviewed By: ryu2

fbshipit-source-id: f1c9590