Be a little more robust to badly-behaved genrules
Summary:
Came across a genrule that did `dirname $OUT` and proceeded to write stuff there.
This is evil, and resulted in build artifacts from a previous build polluting
the next build, because genrules only delete `$OUT` and not the directory containing
`$OUT`. There's no reason not to just nuke the enclosing directory, however, so
might as well.

Test Plan: New integration test

Reviewed By: asp2insp

fbshipit-source-id: 9fa6de5