Add exo-for-resources runtime
Summary:
This adds the runtime part of exo-for-resources. As we don't have build
time support, it's not that helpful. But, you could feasibly manually
rip an apk apart and push the parts and run it with this runtime.

Current downside:
Unknown support for Android 4.0 (works on 4.1+)

The expected metadata is something like:

resources <hash1>
assets <hash2>
...
assets <hashn>

and the corresponding directory contents:

/.../exopackage/<package>/resources/metadata.txt
/.../exopackage/<package>/resources/<hash1>.apk
/.../exopackage/<package>/resources/<hash2>.apk
...
/.../exopackage/<package>/resources/<hashn>.apk

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 764f1cb