Make symlinktree default rulekey propagate content hash
Summary:
SymlinkTree needs to propagate changes in it's sources content through
the default rulekey otherwise dependents' rulekeys won't properly
reflect such changes.

For example:
If some genrule has an input that is the output directory of a
symlinktree, that genrule's rulekey needs to change when the content of
the symlinktree's sources change.

The input-based rulekey continues to ignore the content of non-hashable
source paths so that the symlinktree still gets the benefit of that.

#accept2ship

Test Plan: added test + CI

Reviewed By: plamenko

fbshipit-source-id: 84846c4