Add ForwardingBuildTargetSourcePath which delegates to another SourcePath
Summary: Where a BuildRule wraps another BuildRule and returns the output generated by the thing it wraps. it should return that same SourcePath. However, sometimes BuildTargetSourcePaths are used to detect dependencies, so we should wrap that SourcePath adding ourself as the BuildTarget, so that the dependency on the wrapper is properly detected. Enter: ForwardingBuildTargetSourcePath.

Test Plan: Added tests

Reviewed By: plamenko

fbshipit-source-id: e9a56c6