fix pch rule keys
Summary:
Solved the problem of rule keys not being deterministic, e.g. a rule in one repo, and a rule in another repo, which is in a different fs directory but still the same exact file content (same repo checkout, just in different directories), rule keys can differ.

andrewjcg discovered the root cause, which was that if a rule A requires a PCH, and rule B requires the same PCH, then whichever rule (A or B) "won the race" to instantiating the PCH first, left traces of its preprocessor deps in the PCH rule's key.  This caused lots of cache misses for these keys and thus for lots of other builds.

Test Plan: Developed a repro which catches the problem; the source changes fix those tests.

Reviewed By: andrewjcg

fbshipit-source-id: 02eb7a4