Add `js_library` rule
Summary:
Port the `js_library` rule from build defs to Java.

This rule is used to transpile a set of JavaScript source files with a `worker_tool`. Each `js_library` rule dynamically creates `js_file` rules for each file it contains, and propagates flavors and `extra_args`.

Flavors are currently hardcoded, but will be configurable eventually, with the notable exception of `prod`.

#accept2ship

Test Plan: Integration test with a lot of scenarios

Reviewed By: marcinkosiba

fbshipit-source-id: 48a618a