support macros in ocaml_* compiler_flags
Summary:
allows using `$(exe)` and `$(location)` macros in the `compiler_flags` arg of `ocaml_library` and `ocaml_binary`.

this is a low-level way to enable ppx extensions, by doing

```
ocaml_library(
  ...,
  compiler_flags=['-ppx', '$(exe :my_ppx_rewriter)'],
)
```

this could potentially also be useful for doing `-I $(location ...)`, so I think this is worth doing on its own and also to experiment with ppx support, but we might also want to consider adding `preprocessors=[':my_ppx_rewriter']` support in the future.

Test Plan: unit test

Reviewed By: andrewjcg, illicitonion

fbshipit-source-id: 59f2db2