Allow `TreeBackedElement`s to fall back to their `javac` equivalents when safe
Summary:
A while back we shifted the plan from implementing our own symbol tables to piggybacking
on `javac`'s, particularly for its built-in `ErrorType` support. We have the ability to
generate ABIs from `Element`s now, provided that we have the full classpath available.
This series of commits returns to the problem of resolving types when we don't have all
dependencies, beginning with reworking the previously-written code to fit with the new
plan.

This commit updates `TreeBackedElement` to use the underlying `javac` `Element` when
it is safe to do so, making the necessary changes to the supporting classes.

Test Plan: Existing unit tests

Reviewed By: dreiss

fbshipit-source-id: 71266e4