Remove assets from AaptPackageResources
Summary:
aapt doesn't actually need the assets to package resources (and we
already package some assets in AndroidBinary).

This adds a new BuildRule that merges all the assets that were being
passed to aapt into a single zip which is then added to the final apk
by AndroidBinary.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 1cc2e86