Use JavacTask.enter() instead of analyze
Summary:
`enter` is a private method that does just the Enter phase of compilation.
Previously we were using `analyze`, which goes a little further than that.
That didn't matter before, but now that we're going to back the `TreeBackedElement`s
with both a `Tree` and an `Element`, it matters.

Test Plan: Existing unit tests

Reviewed By: dreiss

fbshipit-source-id: 3fb7e39