Provide `--config client.skip-action-graph-cache=true` option to avoid replacing the cached ActionGraph.
Summary:
Other tools, like Nuclide, may call into Buck to build a compilation database
in order to provide Clang support for the user. Assuming the user is repeatedly
running `buck build <thing-they-are-working-on>`, then the intermittent
`buck build <compilation-database>` from Nuclide will invalidate the action
graph that buckd has in memory, causing the subsequent
`buck build <thing-they-are-working-on>` call to be noticeably slower because
the action graph has to be rebuilt.

Test Plan:
* Repeatedly built an expensive artifact with buckd running.
* Grep the logs and verify we see `ActionGraph cache hit.`
* Build some unrelated thing with `--config client.skip-action-graph-cache=true`.
* Build the expensive artifact again.

**Before this change:**
Would see `ActionGraph cache miss. TargetGraphs mismatched.` twice.

**With this change:**
See `ActionGraph cache miss. TargetGraphs mismatched.` in the logs when building
the unrelated thing, but then see `ActionGraph cache hit.` with the subsequent
build, which is for the expensive artifact. The build time for the expensive
artifact appears to be the same as it would have been without the build of the
unrelated thing inbetween.

Also created `ActionGraphCacheIntegrationTest` to verify this behavior.

Reviewed By: illicitonion

fbshipit-source-id: 5262443