C/C++ library: use metadata to cache/build preprocessor input
Summary:
Currently, we build the exact same exported preprocessor input for C/C++
library rules in both the library interface, when using it for compiling
dependents, and in the description, when using it to compile the library
itself.  Although we try to detect when the library is present to avoid
building this twice, it still can happen in normal flows.

This diff moves this to be built using the metadata framework, so that we
guarantee it's only built once (and therefore, can be hit in the rule key
cache more often).

Test Plan: CI

fbshipit-source-id: 0f50c53