Decouple `TreeBackedTrees` from element creation
Summary:
Earlier I had moved element creation into `TreeBackedTrees`, on the grounds
that it was needed to populate some of the members of that class. However,
I've since realized that it's simpler for `TreeBackedTrees` to simply use
the map of javac `Element` to `TreeBackedElement` that is maintained by
`TreeBackedElements`.

Test Plan: Existing unit tests

Reviewed By: dreiss

fbshipit-source-id: da7f0aa