Have TypeResolver make use of javac's types
Summary:
`TypeResolver` was previously doing all the resolution manually. Now that we're
running `javac`'s enter phase, we can take advantage of what it's figured out.
As of this commit, we're still expecting all dependencies to be present on the
classpath, but that will change shortly.

Test Plan: Existing unit tests

Reviewed By: dreiss

fbshipit-source-id: c4c2181