Simplify time rendering logic
Summary:
Oh `Optional` misuse :( The worst part is that we can still get `null` instead of `Optional` so we need to check both for null and Optional empty. Sigh... this diff cleans this up a bit, but ideally we should not have Optionals as map values here.
The next step is to make `BuildRuleEvent` contain duration information as I need this on other place and it would be redundant and inefficient to compute the same thing on multiple places.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: 3374e41