Set intellij python config for editing buck parser python files.
Summary:
Adds a new intellij module for buck parser code. This module is a bit strange since the same code is a resource root accessible from the buck module. As such, it does not have its own content root, but instead includes the source code via a dependency.

In order set up the Python SDK to get code indexing on the parser files, create a new python SDK and rename it 'Python 2.7'. Subsequently, undo any changes intellij made to the project files.

If you experience weirdness in intellij the first time you rebase past this commit, close and re-open the project.

Test Plan: Verify that if no python sdk is set up in intellij, intellij does not modify the project files by itself.

Reviewed By: bolinfest

fbshipit-source-id: 5637802