Do not log every single action graph cache invalidation
Summary: When rebasing to the stable (or just changing branches) there are usually a lot of changes. All of these changes do the same thing and there is no point reporting every single one of them. This was prompted by a buck-0.log file that had over 65MB most of it going to the 209.000 lines reporting cache invalidation.

Test Plan: tests

Reviewed By: aiked

fbshipit-source-id: d270481