Change how implicit_includes is handled in buck python code
Summary:
Since the `implicit_includes` field is only used to prevent implicit includes
to be included when processing implicit_includes (or their descendants), it's
clearer for it to simply be a boolean of whether a implicit include is being
processed, rather than the list of includes themselves, which is always
`self.implicit_includes`.

Also tweaked documentation and added some PEP484 style type hints.

Test Plan: CI

Reviewed By: mjpieters

fbshipit-source-id: 3b06cb7