Remove BuckConstants.getTrashDir
Summary:
Migrates from BuckConstants.getTrashDir to BuckPaths.getTrashDir

This necessitated changing the API for the async trash cleaner to provide the path to clean
at invocation time, rather than during construction.

This is part of a move to migrate away from BuckConstants so that we can make
buck-out configurable.

The plan:
* Update all usages of `buck-out` to refer to a `Path`
* Convert items which depend on paths from `BuckConstants` to depend on a `BuckPaths` instance instead (you are here)
* Allow `ProjectFilesystem` to handle an absolute path for `buck-out`
* Allow `BuckConfig` to specify the path for `buck-out`
* Test, iterate, bash head on desk, repeat.

Test Plan: updated tests

Reviewed By: Coneko

fbshipit-source-id: 1e57494