improve synchronization & execution in RuleKeyLoggerListener
Summary:
There was a race condition where we would possibly make several unnecessery submissions to the executor. That's because we accessed size outside of lock. Separate size counter is unnecessary. We should use the same executor to submit the final flush, else we might write to the file in an unorderly fashion. It should be safe to submit and shutdown executor as documentation for `ExecutorService.shutdown` says:

    /*
     * Initiates an orderly shutdown in which previously submitted
     * tasks are executed, but no new tasks will be accepted.
     * Invocation has no additional effect if already shut down.
     *
     * <p>This method does not wait for previously submitted tasks to
     * complete execution.  Use {link #awaitTermination awaitTermination}
     * to do that.
     */

Test Plan: check rule_key_logger.tsv

Reviewed By: illicitonion

fbshipit-source-id: 5beaae7