Require that cells can only declare a subset of cell mappings that the root cell declares
Summary:
The fact that cells cannot be uniquely identified by name introduces a number
of inconveniences.  This diff adds a restriction that all cell mappings in
child cells must also be declared in the root cell.

With this restriction in place, we can remove all the complexity of dealing
with cell name resolution contexts and have one-to-one cell name to cell
mappings. This will happen in future diffs.

Also for the future, child cells will only need to declare that it *requires*
the presence of a cell mapping, without having to specify the cell mapping
itself, since the mapping must be what is set in the root cell.

Test Plan:
Added a test for the validation. CI should catch any cases where we actually
have inconsistent namings.

Reviewed By: mzlee

fbshipit-source-id: cfcccb9