Add support to set path prefixes to `JsLibrary`
Summary:
Adds support to set path prefixes for the values used as the `--filename` flag passed to the worker when building a `js_file` rule.

The worker will use the value of that option as the canonical path to the respective file. That way, we can decouple location on disk from virtual location for the packager.

This feature is necessary to install `node_modules` into `buck-out` as a build step, but to allow the packager to treat them as if they would be located in the source tree.

Test Plan: new tests

Reviewed By: marcinkosiba

fbshipit-source-id: 863d523