`JsFile`/`JsLibrary`: allow `Pair<SourcePath, String>` as src
Summary:
In lieu of an implementation of `BuildTargetSourcePath` with support for sub paths, this workaround allows us to use sub paths of build targets as inputs to our rules.

If rebuilding every depending target when anything changes in the upstream dependency turns out to be too expensive, we can still implement this concept as a specialized implementation of `SourcePath`.

Test Plan: new unit & integration tests

Reviewed By: Coneko

fbshipit-source-id: 6e19624