Switch from SDMMobileDevice to FBSimulatorControl for iOS device installation
Summary:
We currently use a third party helper program, SDMMobileDevice to handle the actual loading of the .app file to an iOS device.  This has bit-rotted a bit and has issues with the latest MacOS and iOS versions.

Switch to FBSimulatorControl (which also supports devices, despite the name).  This is owned by us and actively maintained, unlike the former.

Test Plan:
With a wildcard provisioning profile installed, and a device connected:

```
ant test -Dtest.class=InstallCommandIntegrationTest
```

`DemoApp` should be built and installed on the device without any errors.

Reviewed By: Coneko

fbshipit-source-id: c3d7ff2