Reduce DefaultJavaLibrary creation copypasta: Should-be-common Arg fields
Summary:
The main purpose of this commit is to continue reducing boilerplate involved
in `DefaultJavaLibrary` creation. This particular set of changes results in
additional features being enabled, so it's separated from the rest.

Several rules that are derived from `java_library` were being inconsistent about
which `Arg` fields get passed through to the rule. This commit fixes much of that:

- groovy/kotlin/scala rules now honor `proguard_config`
- scala rules and java/robolectric/kotlin test rules now honor `remove_classes`
- all rules now honor `postprocess_classes_commands` (previously just `java_library` did)
- all test rules now honor `provided_deps`

This has a side effect that some nonsensical arguments will get propagated to the
rule that were not previously (specifically test rules can have values for `tests`
and `proguard_config`). This is benign, though in the future we should probably consider
failing when nonsensical args are provided rather than ignoring them.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 50198d3