Don't track requests for files on annotation processor path
Summary:
It was found unnecessary that Buck was including things on the annotation processor path in the dep file. Since the full jars of annotation processor dependencies are also part of the rule key, any changes that affect those jars will trigger a rebuild anyway. So we can remove tracking on things on the annotation processor path, and let it just get the input-based treatment.

Changes:
- `ClassUsageTracker` only wraps requests with trackers if the location is not on the annotation processor path

Test Plan:
- Added new tests that verify that when the location passed in for a request is `ANNOTATION_PROCESSOR_PATH` that those files are not read
- Performed local builds, grabbed the `used-classes.json` file that was being problematic. Confirmed that the entries that should not be in the file are not present (and ONLY those entries).

Reviewed By: jkeljo

fbshipit-source-id: 7f15fd6