Add more test coverage for compiling against ABI JARs
Summary: This diff adds some more coverage in compiling against ABI JARs. There are a lot of cases and it may not be practical to have tests that compile against ABIs for every single case. Instead, the tests added are more towards testing possible problematic cases and also cases that have shown to be problematic in the past.

Test Plan: Ran the added tests to ensure they work

Reviewed By: jkeljo

fbshipit-source-id: 998a6aa