Don't create empty intellij modules for cross-cell dependencies
Summary:
Project file generation isn't cell-aware. Before this change, we would
find all cross-cell dependencies, and create a module for them, but the
module was guaranteed to be empty because it tried to resolve the files
in it relative to the root cell's root, not its cell's. This has the
side-effect of creating a shadow structure of the other cell's
filesystem in the root cell, where the only files present are .iml
files, which don't reference any actual files.

Instead, explicitly filter out other cells when generating modules. This
means that in the non-project buck code (e.g. the ProjectFilesystem), we
don't need to worry about this broken cross-cell behaviour where we try
to look up non-existent files.

Reviewed By: mzlee

fbshipit-source-id: 24927cc