set doNotify if a thread is going to wait for buck command
Summary:
If a thread is going to wait for buck command process, it will wait
for a semaphore, but the semaphore will be released when buck command process
terminates only if doNotify is set.  Without this, the wait never completes.

Test Plan: unit test and verify with manual test inside intellij

Reviewed By: styurin

fbshipit-source-id: 4b0a07c