Create a BuildInfoStore for metadata
Summary:
We want to store build metadata outside of the filesystem (i.e., in a database)
to speed up large builds that have to write a lot of metadata.  This diff
introduces a BuildInfoStore object that is responsible for reading and writing
said metadata.  For now, this object simply uses the filesystem, so there should
be no behavioral change.  A later diff will introduce alternative
implementations.

To minimize plumbing, I create BuildInfoStores as needed within
CachingBuildEngine, one per ProjectFilesystem.  This keeps plumbing to a minimum
-- I've routed the store into BuildInfoRecorder and OnDiskBuildInfo since those
currently deal with reading and writing the metadata, and it seems like more
churn to change that.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 358e744