Disable autodeps command.
Summary:
After extended discussions the team has decided not to continue using
the autodeps command. This disables it with an error message and a
setting to reenable it to ease the transition.

The autodeps command is based on various assumptions and heuristics
that only hold true in Buck's codebase, as such it should not be a built in
command.

Note this does not disable the `BUCK.autodeps` auxiliary file mechanism.
There are currently no plans for removing that.

Test Plan:
CI
run `buck autodeps` see error message

Reviewed By: illicitonion, aiked

fbshipit-source-id: bf59a38