Avoid unnecessary deletions while unzipping artifacts
Summary:
When running a build that updates lots of artifacts, buck spends a lot
of time preemptively removing directories, only to recreate them.  We can avoid
this overhead by tracking which files currently exist, removing those which are
about to be overwritten, and only deleting the remainder.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 93e9975