Build symbol table in a TaskListener rather than the JavacTask itself
Summary:
The compiler rebuilds the symbol table after each round of annotation processing. We can't keep
up with that if we're just handling the `enter` method of `JavacTask`.

Test Plan:
Code inspection for the new `clear` methods. When we introduce annotation processor support
any cases of not clearing will show up pretty quickly with those tests.

Existing unit tests should detect if we ever don't enter something that we should

Reviewed By: dreiss

fbshipit-source-id: 9f05155