Reduce DefaultJavaLibrary creation copypasta: Get shouldSuggestDependencies from `*BuckConfig`
Summary:
The `Description` classes for non-Java JVM languages already had their
appropriate `BuckConfig`s available, so might as well add `JavaBuckConfig`
to the Java `Descriptions` and use it.

This will also come in handy for the separation of `Javac` and `JavacOptions`
project that I'm working on right now.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 343bc97