RocksDB-based BuildInfoStore
Summary:
Import RocksDB as a third-party jar and use it to store build metadata.

Keys in the DB are of the form '//fully/qualified:target\0METADATA_KEY'.
Deletes are done by iterating over the "target" space; writes are batched, while
reads are point queries (simply because this is the way the APIs wrapping
BuildInfoStore were already set up).

The metadata storage engine is selectable via:
```
--config build.metadata_storage={filesystem,rocksdb}
```

If for some reason the RocksDB database can't be opened (whether because there's
no shared library for the platform, or because another process already has the
DB locked) then we fall back to FilesystemBuildInfoStore.

Test Plan:
```
./scripts/build-perf/build_helper.py -n3 scripts/build-perf/scenarios/slurp
```
shows somewhere between 0 and 20 seconds improvement on a ~120 second build.

Reviewed By: dreiss

fbshipit-source-id: db909c7