Make RuleKeyFactory's builders generic
Summary:
In a series of previous refactors, base `RuleKeyBuilder` was made generic to support arbitrary `RuleKeyHashers`. However, concrete implementations in rule key factories were still left hardcoded to a particular hasher. This diff makes concrete `RuleKeyBuilder` implementations be generic as well.

This will allow us to generate diagnostics rule keys with a diagnostics hasher implementation (e.g. string or thrift hasher) only for the rules we want to (e.g. those that were locally built), as opposed to all rules (which is how this is presently implemented when verbose rule key logging is enabled).

The scope of components involved:
* `RuleKeyHasher` - low level implementation of how Buck and Java primitives (e.g. String, BuildTarget, ...) get converted to a particular hash (e.g SHA1, String, ...). This class is "dumb" in a sense that it doesn't interpret objects it hashes, but just serializes them as is.
* `RuleKeyBuilder` - performs a recursive traversal over a single build rule's field that should be included in a rule key and feeds its components to the underlying `RuleKeyHasher`. Unlike hasher, `RuleKeyBuilder` interprets objects. For example, list-like objects get traversed and each element gets processed recursively. `BuildRule` and `SourcePath` objects usually get some special handling too.
* `RuleKeyFactory` - Provides a specialized `RuleKeyBuilder` implementation in accordance with this factory's function (default, dep-file, input-based, ...). Feeds its builder with the build rule's fields (obtained with help of `RuleKeyFieldLoader`). Optionally performs caching on per `BuildRule` and `RuleKeyAppendable` boundaries.

Test Plan: tests

Reviewed By: illicitonion

fbshipit-source-id: b1aaa77