Remove an extra space added when subcommand has no arguments in ChromeTraceParser.
Summary:
Small fix so the value returned from `ChromeTraceParser.COMMAND` is consistent:
now it never has an extra space at the end.

Test Plan:
I have some downstream changes that were testing this value and I was surprised
to find an extra space at the end, so this fixes that.

Reviewed By: Coneko

fbshipit-source-id: 463987c