Remove additionalClasspathEntries
Summary:
The only situation `additionalClasspathEntries` for `DefaultJavaLibrary` was being used was when `AndroidLibraryDescription` had a `DummyRDotJava`, a build rule, that needed to be a part of the classpath for compilation. The reason why this was separated from the rest of the dependencies was due to `DefaultJavaLibrary` filtering dependencies for the classpath that do not implement `JavaLibrary`.

These changes remove the additionalClasspathEntries parameter to the constructor of DefaultJavaLibrary and anything extending it completely. To still be able to add the DummyRDotJava rule to the classpath, `DefaultJavaLibrary` filters by `HasJavaAbi` instead, which is only used by classes extending `JavaLibrary` anyway, with the exception of `DummyRDotJava`.

Going forward we will be changing `DefaultJavaLibrary` to compile against ABIs. Additional clean up will be performed when that switch is made, since we'll be able to just take the deps in `abiClasspath` and supply that to the classpath.

Test Plan: Ran tests to ensure nothing broke from these changes, and tested locally by building Android libraries that include DummyRDotJava build rules

Reviewed By: AsyncDBConnMarkedDownDBException

fbshipit-source-id: f82aa33