Create a default private constructor in a class if needed
Summary:
We don't stub private constructor methods, but if stripping all of these constructors results in no non-synthetic constructors, we want to at least include a default private constructor. This is because removing all these private methods will leave the class looking like it has no constructors at all, which is not possible. A default constructor is generated by the compiler and would appear in the bytecode.

With this diff, we will track if the class we are visiting has any public, non-synthetic constructor, and is not an interface or annotation. If we do not find this type of constructor, then we will generate a default private constructor in `AbiFilteringClassVisitor.visitEnd()`. `ClassVisitorDriverFromElement.visitExecutable()` had to be modified to pass private methods to the `ClassVisitor` so that `AbiFilteringClassVisitor` can handle this case.

Test Plan: New tests have been added to `StubJarTest` and `StubJarTest.assertMembersStubbedCorrectly()` has been modified so it handles the new logic. Existing tests help provide coverage as well. Also put the "compile against ABIs" changes on top and verified that the previously found bug in compiling against ABIs is fixed and that compiling is still successful.

Reviewed By: jkeljo

fbshipit-source-id: 789a6f4