Separate Javac and JavacOptions: Compute transitive imports at instantiation time
Summary:
Previously a `Javac` instance specified only its first level classpath. There was
code in `JavacOptions` that computed the transitive classpath so that the necessary
rule dependencies could be created. There was *different* code in `JavacStep` that
computed the classpath that was actually used to run the compiler.

The main goal of this commit is to put the full transitive classpath in the
`Javac` instance itself, so that `JavacOptions` logic can just ask the instance.
It accomplishes this by consolidating the two slightly different classpath
computations into `JavacProvider` and using the same classpath for both
dependency computation and runtime.

Test Plan: Unit tests, JarBackedJavac integration test

Reviewed By: dreiss

fbshipit-source-id: d40a73b