Add code sign entitlements to generated Xcode project
Summary:
When building using a Xcode generated project for iOS devices, Xcode's Capabilities tab doesn't show the right capabilities from the entitlements file that we sign with. Even though Buck sets the right file in the Build Settings tab, the UI still shows an incomplete list. And if you make any changes here, Xcode creates a brand new entitlements file, completely ignoring the file that it should be using during the code signing phase.

Turns out the problem here is because Buck has not added our entitlements file to the project itself. After manually adding the entitlements plist to the project, the UI correctly represents the capabilities that the app requests and future modifications to that UI will update the correct file rather than creating a new file.

Test Plan:
* Unit tests
* Opened generated Xcode project with `CODE_SIGN_ENTITLEMENTS` defined, observe entitlements plist in the Xcode project structure.  Modifying "capabilities" causes Xcode to modify the referenced entitlements file properly.

Reviewed By: dinhviethoa

fbshipit-source-id: 53d5b40