Add a debug helper in ZipScrubber
Summary:
When debugging ZipScrubber, it's useful to be able to get the name of
the current entry, but it seems like IntelliJ's "watch" window isn't
able to watch a set of statements.  Pull the logic out into a helper
function ot make it easier to use.

Test Plan: Used it in the debugger.

Reviewed By: illicitonion

fbshipit-source-id: 7ea3af1