Make ZipScrubber fix aapt2 output for ZipInputStream
Summary:
ZipInputStream doesn't use the zip central directory.  It just processes
the zip file as a sequence of "local file entry + file content".
Therefore, it can't handle non-deflated files that don't have a size in
their local file entry.

Unfortunately, "aapt2 link" produces exactly one entry like this:
resources.arsc.  We can use ZipScrubber to copy the information from the
central directory to the local file entry.  The current in-place
structure of ZipScrubber doesn't allow us to remove the data descriptor
from after the file, so ZipInputStream will treat this as the last file
in the zip.  Fortunately, resources.arsc is already the last file in the
zip.

In short, this won't hurt any zip files, but it also won't help if we
encounter one where the stored file isn't the last entry.  If we end up
in that situation, we can make this code more robust.

Test Plan: Built with aapt2.

Reviewed By: Coneko

fbshipit-source-id: b41adf0