aapt2: Stop using the AaptPackageResources rule as a data object
Summary:
Previously, AaptPackageResources (the rule that runs aapt) had a bunch
of public methods for exposing paths that were needed by downstream
rules.  This was in the way of enabling aapt2 (which won't use
AaptPackageResources).

Instead, create an immutable object to wrap up all of this data, and use
it where necessary in AndroidBinaryGraphEnhancer.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 8328326