Add ExopackageInstaller integration tests
Summary:
This moves the device interaction stuff in ExopackageInstaller into an
inner class. The tests then simulate the behavior of that interaction.

Adds tests that, given various ExopackageInfo and differences in the
files referenced in that info, the ExopackageInstaller gets the device
into the correct state (i.e. apk is installed, dexes/libs/etc are
pushed as expected) and that it only installs as many files as expected
(e.g. a dex change only requires installing that dex).

With some changes, we could use this device simulation to
actually verify much of the behavior of full exopackage builds (though
not that the installed app actually works correctly at runtime).

Test Plan: Added test

Reviewed By: dreiss

fbshipit-source-id: 7ad6b45