Explicitly create ThreadPoolExecutor used by OkHttpClient and close it on close
Summary:
ClientSideSlb's OkHttpClient's internal Executor was never closed, leading to
threads that may hang around for up to 60 seconds.

This change makes the thread pool's creation explicit, and give the threads the
same priority as the ping scheduler thread and a more descriptive name.

Test Plan:
run buck query commands with the daemon a few times, see that the
threads with this name are closed after the commands are run using a profiler.

Reviewed By: marcinkosiba

fbshipit-source-id: 2e07000