Remove assets from AaptPackageResources (third time's the charm).
Summary:
aapt doesn't actually need the assets to package resources (and we
already package some assets in AndroidBinary).

This adds a new BuildRule that merges all the assets and the aapt output
into a single zip.

This rejects any files of the form *.gz. Old aapt handle these
differently than aapt2, and it would be confusing to use one or the
other while transitioning.

Test Plan:
Added tests:
 1. app contains assets from android_resource
 2. changing asset triggers rebuild correctly
 3. app assets are compressed
 4. *.gz files cause build failure

Reviewed By: dreiss

fbshipit-source-id: 5598b9b