Introduce FocusedModuleTargetMatcher instead of using Optional<ImmutableSet<...>>
Summary:
Instead of adding a special API on BuildTarget specifically to support the
"absent-means-match-everything" used for matching stuff for focus, put this
logic in a little class.

The matcher class encapsulates the Predicate nature of the test of whether a
target is focused on, in addition to the optional nature of the parameter.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: 91eef9f