Add a debounced 5 second delay to after-command GC
Summary:
GC is run after a command to return memory to the system after long running
commands.  If commands are invoked in quick succession, such as for queries
invoked by scripts, the full GC may affect the latency of the next command
being run.

This change schedules the after-command GC to happen after 5 seconds after the
last command is run. If another command is run within that interval, the GC is
further pushed back.

Test Plan:
Run commands in quick succession, see that gc does not trigger via
VM telemetry.

Reviewed By: illicitonion

fbshipit-source-id: 43f215c